+++ Classes/Private/Manager/Submanagers/OCTSubmanagerChatsImpl.m
@@ -155,7 +155,9 @@
     NSPredicate *predicate = [NSPredicate predicateWithFormat:@"chatUniqueIdentifier == %@ AND messageText.messageId == %d",
                               chat.uniqueIdentifier, messageId];
 
+    // messageId is reset on every launch, so we want to update delivered status on latest message.
     RLMResults *results = [realmManager objectsWithClass:[OCTMessageAbstract class] predicate:predicate];
+    results = [results sortedResultsUsingProperty:@"dateInterval" ascending:NO];
 
     OCTMessageAbstract *message = [results firstObject];